PhastJustHelper addds deleteFaultyPhasenJustageItem to PhaseJustage
we found that there are values that do not match our expectations e.g.: 19 times 0.0343149 but one : 1.72088
this will lead to Justgae issue! solving this might be done:
 1) by finding the root cause of the issue
 2) by repeating the measurement
 3) by deleting the faulty measurement
alldough the first would be the right solution and the second is a right solution, we are going to use the third approach!

Signed-off-by: Wilfried Schunk <w.schunk@zera.de>